+++ src/main/java/com/github/unchama/seichiassist/data/PlayerData.kt
@@ -8,7 +8,7 @@ import com.github.unchama.seichiassist.minestack.MineStackHistoryData
 import com.github.unchama.seichiassist.minestack.MineStackObj
 import com.github.unchama.seichiassist.task.MebiusTask
 import com.github.unchama.seichiassist.task.VotingFairyTask
-import com.github.unchama.seichiassist.util.ExperienceManager
+import com.github.unchama.seichiassist.util.exp.ExperienceManager
 import com.github.unchama.seichiassist.util.Util
 import com.github.unchama.seichiassist.util.Util.DirectionType
 import org.bukkit.*
